feat(enrollment): add `EnrollmentSummariesWidget` and `FirstRegistrationPage`, refactor `PreRegistrationsPage`
- Introduced `EnrollmentSummariesWidget` for reusable enrollment summary handling.
- Added `FirstRegistrationPage` and integrated it into the navigation structure.
- Refactored `PreRegistrationsPage` to use `EnrollmentSummariesWidget` to improve maintainability and simplify logic.
- Enhanced `PreRegistrationsInfoBar` with dynamic badge and action support.
