+++ lib/features/home/presentation/pages/home_page.dart
@@ -2,7 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:school_app_flutter/core/constants/menu_constants.dart';
 import 'package:school_app_flutter/core/theme/app_theme.dart';
+import 'package:school_app_flutter/features/enrollment/presentation/context/enrollment_detail_intent.dart';
+import 'package:school_app_flutter/features/enrollment/presentation/context/enrollment_detail_origin.dart';
 import 'package:school_app_flutter/features/enrollment/presentation/pages/enrollment_feature_scope.dart';
+import 'package:school_app_flutter/features/enrollment/presentation/pages/first_registration_page.dart';
 import 'package:school_app_flutter/features/enrollment/presentation/pages/pre_registrations_page.dart';
 import 'package:school_app_flutter/features/enrollment/presentation/pages/re_registrations_page.dart';
 import 'package:school_app_flutter/features/home/presentation/bloc/navigation_bloc.dart';
@@ -78,7 +81,8 @@ class _HomePageView extends StatelessWidget {
   Widget _buildMainContent(BuildContext context, NavigationState state) {
     final hideBreadcrumbForEnrollment =
         state.selectedSubMenuId == MenuConstants.preInscriptionsId ||
-        state.selectedSubMenuId == MenuConstants.reInscriptionsId;
+        state.selectedSubMenuId == MenuConstants.reInscriptionsId ||
+        state.selectedSubMenuId == MenuConstants.premiereInscriptionId;
 
     return Container(
       width: double.infinity,
@@ -152,6 +156,8 @@ class _HomePageView extends StatelessWidget {
         return const EnrollmentFeatureScope(child: PreRegistrationsPage());
       case MenuConstants.reInscriptionsId:
         return const EnrollmentFeatureScope(child: ReRegistrationsPage());
+      case MenuConstants.premiereInscriptionId:
+        return const EnrollmentFeatureScope(child: FirstRegistrationPage());
       default:
         return Container(
           width: double.infinity,
